+++ cli/internal/providers/datacatalog/localcatalog/model.go
