+++ cli/internal/providers/datacatalog/localcatalog/testdata/properties_batch_1.yaml
