+++ cli/internal/providers/datacatalog/localcatalog/testdata/properties_batch_2.yaml
